refactor(parse): Move parseWordDictEntry tests to their own file
